Add httpmock, merge unmarshal into fileutil. (#434)
## Community Contribution License
All community contributions in this pull request are licensed to the
project maintainers
under the terms of the [Apache 2
license](https://www.apache.org/licenses/LICENSE-2.0).
By creating this pull request I represent that I have the right to
license the
contributions to the project maintainers under the Apache 2 license.

## Summary
Add httpmock as an opensource package. It seems generally helpful, and
we'll need it for aisdk.
Decided to merge `unmarshal` into `fileutil` .... but let me know if you
disagree. In the process, added better docs, tests, and added a Glob
function to fileutil.

## How was it tested?
Ran tests.