Implement `serror` a library for structured errors (#441)
## Summary
Implement `serror` a library for structured errors

## How was it tested?
devbox run test

## Community Contribution License
All community contributions in this pull request are licensed to the
project maintainers
under the terms of the [Apache 2
license](https://www.apache.org/licenses/LICENSE-2.0).
By creating this pull request I represent that I have the right to
license the
contributions to the project maintainers under the Apache 2 license.