Rename package `result` to `try` (#455)
## Summary
After using `result.Result` for a bit to handle wrapping values and
errors, I feel like the name `Result` is too generic and often collides
with what the user of the package might be doing. I ran into an example
where people were calling this type of wrapper `Try[T]`, and I liked it.
This PR renames `Result` to `Try`. I tweaked the names of a few
functions as well.

## How was it tested?
Ran the tests.

Once this is checked in, I will update axiom to use the latest naming.

## Community Contribution License

All community contributions in this pull request are licensed to the
project
maintainers under the terms of the
[Apache 2 License](https://www.apache.org/licenses/LICENSE-2.0).

By creating this pull request I represent that I have the right to
license the
contributions to the project maintainers under the Apache 2 License as
stated in
the
[Community Contribution
License](https://github.com/jetify-com/opensource/blob/main/CONTRIBUTING.md#community-contribution-license).