Add ReplayServer (#462)
## Summary
Adds a new type of server to `httpmock`: a ReplayServer. This type of
server allows you to talk to a real API when first creating the test and
then records the interaction for replay later.

## How was it tested?
Unit tests.

## Community Contribution License

All community contributions in this pull request are licensed to the
project maintainers under the terms of the [Apache 2
License](https://www.apache.org/licenses/LICENSE-2.0).

By creating this pull request I represent that I have the right to
license the contributions to the project maintainers under the Apache 2
License as stated in the [Community Contribution
License](https://github.com/jetify-com/opensource/blob/main/CONTRIBUTING.md#community-contribution-license).