Add gzip support to httpmock (#554)
## Summary
When using httpmock with gzip compression the data was being saved in a
corrupted state. This PR adds a gzip hook that automatically
decompresses data before it's stored. It also makes the data easier to
inspect.

## How was it tested?
Wrote unit test.

## Community Contribution License

All community contributions in this pull request are licensed to the
project maintainers under the terms of the [Apache 2
License](https://www.apache.org/licenses/LICENSE-2.0).

By creating this pull request I represent that I have the right to
license the contributions to the project maintainers under the Apache 2
License as stated in the [Community Contribution
License](https://github.com/jetify-com/opensource/blob/main/CONTRIBUTING.md#community-contribution-license).